Add snapshotter.type flag to build-disk command
When building a disk image we do not specify the snapshotter, which
configures the expansion mechanism to set COS_STATE partition filesystem
to ext4.

The green flavor uses the btrfs snapshotter, which means when it boots
and tries to mount the COS_STATE partition it expects a btrfs
filesystem, but finds an ext4 one, which errors.

In this commit we add the --snapshotter.type flag to the build-disk
command and set the snapshotter in the Makefile accordingly.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>